fixed major bug (61 seconds exist right?)

+++ src/sleep.h
@@ -19,7 +19,7 @@
 
 void sleep_goto_sleep_until(datetime_t *t_alarm, rtc_callback_t callback, bool xosc_en = true);
 
-void sleep_goto_sleep_for(long seconds, rtc_callback_t callback, bool xosc_en = true);
+void sleep_goto_sleep_for(uint32_t seconds, rtc_callback_t callback, bool xosc_en = true);
